pyproject.toml
[project]
name = "workspace-colors"
version = "0.1.0"
description = "Color picker for VS Code workspace themes"
requires-python = ">=3.10"
dependencies = [
    "matplotlib>=3.10.8",
    "numpy>=2.2.6",
    "pillow>=12.0.0",
    "textual>=0.40.0",
]

[project.scripts]
colorpick = "color_picker.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["color_picker"]

[tool.hatch.build.targets.wheel.sources]
"color_picker" = "color_picker"

[tool.hatch.build]
include = [
    "color_picker/*.py",
    "color_picker/colors.json",
]
